(ID13) Fixed bugs ID1308vv.sma, ID1310vv.kya

+++ src/components/activeParametersEditors/IntegerTextEditor.js
@@ -7,7 +7,7 @@ export default function IntegerTextEditor(props) {
 
     return (
         <NumericTextBox className='parametereditor'
-            value={parseInt(value)}
+            value={parseFloat(value)}
             name={props.id}
             onChange={props.selectionChanged}
         />
